Logically dead code: CID 52756 Nesting level does not match identation: CID 53021 Unitialized scalar fields: CID 53236 Commented out unused function CID 107031 Printf arg type mismatch: CID 52692

+++ src/world/MapMgr.h
@@ -252,7 +252,7 @@ class SERVER_DECL MapMgr : public CellHandler <MapCell>, public EventableObject,
 		virtual int32 event_GetInstanceID() { return m_instanceID; }
 
 		void LoadAllCells();
-		size_t GetPlayerCount() { return m_PlayerStorage.size(); }
+		uint32 GetPlayerCount() { return m_PlayerStorage.size(); }
 		uint32 GetTeamPlayersCount(uint32 teamId);
 
 		void _PerformObjectDuties();
